completed JWT authenication

+++ controllers/api/v1/posts_api.js
@@ -19,20 +19,24 @@ module.exports.index=async function(req,res){
 module.exports.destroy = async (req, res) => {
     try {
       let post = await Post.findById(req.params.id);
-      //.id means converting object into string
-      //if ((post.user = req.user.id)) {
+      if ((post.user = req.user.id)) {
         post.remove();
         await Comment.deleteMany({ post: req.params.id });
         
         //req.flash("success", "Post deleted !!!");
-        return res.json(200,{
+        return res.status(200).json({
             message:'Post deleted'
         });
+      } else {
+        return res.status(401).json({
+          message: "You are not authorized to delete this post",
+        });
+      }
       
     } catch (err) {
       //req.flash("error", "Error in deleting post");
 
-      return res.json(500,{
+      return res.status(500).json({
         message:'Error in deleting post',
       });
     }
